+++ cmd/provenanced/cmd/simulate_test.go
@@ -6,8 +6,6 @@ import (
 	"path/filepath"
 	"testing"
 
-	"github.com/provenance-io/provenance/internal/antewrapper"
-
 	"github.com/cosmos/cosmos-sdk/client/flags"
 	"github.com/cosmos/cosmos-sdk/crypto/keys/secp256k1"
 	clitestutil "github.com/cosmos/cosmos-sdk/testutil/cli"
@@ -19,6 +17,7 @@ import (
 	"github.com/stretchr/testify/suite"
 
 	"github.com/provenance-io/provenance/cmd/provenanced/cmd"
+	"github.com/provenance-io/provenance/internal/antewrapper"
 	"github.com/provenance-io/provenance/testutil"
 	msgfeestypes "github.com/provenance-io/provenance/x/msgfees/types"
 )
@@ -82,6 +81,10 @@ func (s *SimulateTestSuite) SetupTest() {
 
 	_, err = s.testnet.WaitForHeight(1)
 	s.Require().NoError(err)
+}
+
+func (s *SimulateTestSuite) TearDownTest() {
+	testutil.CleanUp(s.testnet, s.T())
 }
 
 func TestSimulateTestSuite(t *testing.T) {