+++ CHANGELOG.md
@@ -42,6 +42,8 @@ Ref: https://keepachangelog.com/en/1.0.0/
 * Add `add-genesis-msg-fee` command to add msg fees to genesis.json and update Makefile to have pre-defined msg fees [#667](https://github.com/provenance-io/provenance/issues/667)
 * Add msgfees summary event to be emitted when there are txs that have fees [#678](https://github.com/provenance-io/provenance/issues/678)
 * Adds home subcommand to the cli's config command [#620] (https://github.com/provenance-io/provenance/issues/620)
+* Add support for rocksdb and badgerdb [#702](https://github.com/provenance-io/provenance/issues/702)
+* Create `dbmigrate` utility for migrating a data folder to use a different db backend [#696](https://github.com/provenance-io/provenance/issues/696)
 
 ### Improvements